Completed implementation of GzipRequestCustomizer
