+++ jetty-server/src/main/config/modules/gzip-inflate.mod
@@ -1,0 +1,17 @@
+#
+# GZIP inflate module
+# Applies GzipRequestCustomizer to entire server to inflate gzipped requests
+#
+
+[depend]
+server
+
+[xml]
+etc/jetty-gzip-inflate.xml
+
+[ini-template]
+## Buffer size for compressed data
+# jetty.gzip.inflate.compressedBufferSize=4096
+
+## Buffer size for compressed data
+# jetty.gzip.inflate.inflatedBufferSize=16384
